[FrameworkBundle] remove deprecated code paths
