+++ src/Symfony/Bundle/FrameworkBundle/Command/AssetsInstallCommand.php
@@ -44,13 +44,9 @@ class AssetsInstallCommand extends Command
     private $filesystem;
     private $projectDir;
 
-    public function __construct(Filesystem $filesystem, string $projectDir = null)
+    public function __construct(Filesystem $filesystem, string $projectDir)
     {
         parent::__construct();
-
-        if (null === $projectDir) {
-            @trigger_error(sprintf('Not passing the project directory to the constructor of %s is deprecated since Symfony 4.3 and will not be supported in 5.0.', __CLASS__), E_USER_DEPRECATED);
-        }
 
         $this->filesystem = $filesystem;
         $this->projectDir = $projectDir;
