+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/identity_translator.xml
@@ -11,8 +11,5 @@
         <service id="Symfony\Contracts\Translation\TranslatorInterface" alias="translator" />
 
         <service id="identity_translator" class="Symfony\Component\Translation\IdentityTranslator" />
-        <service id="translator.selector" class="Symfony\Component\Translation\MessageSelector">
-            <deprecated>The "%service_id%" service is deprecated since Symfony 4.2, use "identity_translator" instead.</deprecated>
-        </service>
     </services>
 </container>
