+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/session.xml
@@ -64,10 +64,6 @@
             </argument>
         </service>
 
-        <service id="session.save_listener" class="Symfony\Component\HttpKernel\EventListener\SaveSessionListener">
-            <deprecated>The "%service_id%" service is deprecated since Symfony 4.1. Use the "session_listener" service instead.</deprecated>
-        </service>
-
         <!-- for BC -->
         <service id="session.storage.filesystem" alias="session.storage.mock_file" />
     </services>
