--- src/Symfony/Bundle/FrameworkBundle/Tests/Controller/TestController.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Symfony\Bundle\FrameworkBundle\Tests\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-class TestController extends Controller
-{
-    use TestControllerTrait;
-}
