--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_strict_email_and_validation_mode.php
@@ -1,8 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'validation' => [
-        'strict_email' => true,
-        'email_validation_mode' => 'strict',
-    ],
-]);
