--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/validation_strict_email_and_validation_mode.yml
@@ -1,4 +1,0 @@
-framework:
-    validation:
-        strict_email: true
-        email_validation_mode: html5
