--- src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/yml/validation_strict_email_disabled.yml
@@ -1,3 +1,0 @@
-framework:
-    validation:
-        strict_email: false
